requirements.txt
soundcard
numpy
openai-whisper
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
argostranslate